FIX: sorting did not work correctly with tree structure.
A parent compared to one of its children would return the same node at the same level as each other that share a common parent.  It would then compare the parent with itself and conclude they are equal, when in fact the child should be lower in the tree than the parent.